Support empty/default contexts, put chrono/regex behind features
