+++ interpreter/Cargo.toml
@@ -11,11 +11,11 @@ categories = ["compilers"]
 [dependencies]
 cel-parser = { path = "../parser", version = "0.7.1 " }
 thiserror = "1.0.40"
-chrono = { version = "0.4.26", default-features = false, features = ["alloc"] }
+chrono = { version = "0.4.26", default-features = false, features = ["alloc"], optional = true }
 nom = "7.1.3"
 paste = "1.0.14"
 serde = "1.0.196"
-regex = "1.10.5"
+regex = { version = "1.10.5", optional = true }
 serde_json = { version = "1.0.124", optional = true }
 base64 = { version = "0.22.1", optional = true }
 
@@ -29,4 +29,5 @@ harness = false
 
 [features]
 json = ["dep:base64", "dep:serde_json"]
-
+regex = ["dep:regex"]
+chrono = ["dep:chrono"]
