+++ interpreter/src/context.rs
@@ -138,6 +138,24 @@ impl<'a> Context<'a> {
             variables: Default::default(),
         }
     }
+
+    /// Constructs a new empty context with no variables or functions.
+    ///
+    /// If you're looking for a context that has all the standard methods, functions
+    /// and macros already added to the context, use [`Context::default`] instead.
+    ///
+    /// # Example
+    /// ```
+    /// use cel_interpreter::Context;
+    /// let mut context = Context::empty();
+    /// context.add_function("add", |a: i64, b: i64| a + b);
+    /// ```
+    pub fn empty() -> Self {
+        Context::Root {
+            variables: Default::default(),
+            functions: Default::default(),
+        }
+    }
 }
 
 impl<'a> Default for Context<'a> {
@@ -146,6 +164,15 @@ impl<'a> Default for Context<'a> {
             variables: Default::default(),
             functions: Default::default(),
         };
+
+        #[cfg(feature = "regex")]
+        ctx.add_function("matches", functions::matches);
+
+        #[cfg(feature = "chrono")]
+        ctx.add_function("duration", functions::duration);
+        #[cfg(feature = "chrono")]
+        ctx.add_function("timestamp", functions::timestamp);
+
         ctx.add_function("contains", functions::contains);
         ctx.add_function("size", functions::size);
         ctx.add_function("has", functions::has);
@@ -155,9 +182,6 @@ impl<'a> Default for Context<'a> {
         ctx.add_function("max", functions::max);
         ctx.add_function("startsWith", functions::starts_with);
         ctx.add_function("endsWith", functions::ends_with);
-        ctx.add_function("matches", functions::matches);
-        ctx.add_function("duration", functions::duration);
-        ctx.add_function("timestamp", functions::timestamp);
         ctx.add_function("string", functions::string);
         ctx.add_function("bytes", functions::bytes);
         ctx.add_function("double", functions::double);
