+++ interpreter/src/functions.rs
@@ -1,11 +1,14 @@
 use crate::context::Context;
+#[cfg(feature = "chrono")]
 use crate::duration::{format_duration, parse_duration};
 use crate::magic::{Arguments, Identifier, This};
 use crate::objects::{Value, ValueType};
 use crate::resolvers::{Argument, Resolver};
 use crate::ExecutionError;
 use cel_parser::Expression;
+#[cfg(feature = "chrono")]
 use chrono::{DateTime, Duration, FixedOffset};
+#[cfg(feature = "regex")]
 use regex::Regex;
 use std::cmp::Ordering;
 use std::convert::TryInto;
@@ -152,7 +155,9 @@ pub fn contains(This(this): This<Value>, arg: Value) -> Result<Value> {
 pub fn string(ftx: &FunctionContext, This(this): This<Value>) -> Result<Value> {
     Ok(match this {
         Value::String(v) => Value::String(v.clone()),
+        #[cfg(feature = "chrono")]
         Value::Timestamp(t) => Value::String(t.to_rfc3339().into()),
+        #[cfg(feature = "chrono")]
         Value::Duration(v) => Value::String(format_duration(&v).into()),
         Value::Int(v) => Value::String(v.to_string().into()),
         Value::UInt(v) => Value::String(v.to_string().into()),
@@ -247,6 +252,7 @@ pub fn ends_with(This(this): This<Arc<String>>, suffix: Arc<String>) -> bool {
 /// ```cel
 /// "abc".matches("^[a-z]*$") == true
 /// ```
+#[cfg(feature = "regex")]
 pub fn matches(
     ftx: &FunctionContext,
     This(this): This<Arc<String>>,
@@ -499,17 +505,34 @@ pub fn exists_one(
 /// - `1.5ms` parses as 1 millisecond and 500 microseconds
 /// - `1ns` parses as 1 nanosecond
 /// - `1.5ns` parses as 1 nanosecond (sub-nanosecond durations not supported)
+#[cfg(feature = "chrono")]
 pub fn duration(value: Arc<String>) -> Result<Value> {
     Ok(Value::Duration(_duration(value.as_str())?))
 }
 
 /// Timestamp parses the provided argument into a [`Value::Timestamp`] value.
 /// The
+#[cfg(feature = "chrono")]
 pub fn timestamp(value: Arc<String>) -> Result<Value> {
     Ok(Value::Timestamp(
         DateTime::parse_from_rfc3339(value.as_str())
             .map_err(|e| ExecutionError::function_error("timestamp", e.to_string().as_str()))?,
     ))
+}
+
+/// A wrapper around [`parse_duration`] that converts errors into [`ExecutionError`].
+/// and only returns the duration, rather than returning the remaining input.
+#[cfg(feature = "chrono")]
+fn _duration(i: &str) -> Result<Duration> {
+    let (_, duration) =
+        parse_duration(i).map_err(|e| ExecutionError::function_error("duration", e.to_string()))?;
+    Ok(duration)
+}
+
+#[cfg(feature = "chrono")]
+fn _timestamp(i: &str) -> Result<DateTime<FixedOffset>> {
+    DateTime::parse_from_rfc3339(i)
+        .map_err(|e| ExecutionError::function_error("timestamp", e.to_string()))
 }
 
 pub fn max(Arguments(args): Arguments) -> Result<Value> {
@@ -536,24 +559,15 @@ pub fn max(Arguments(args): Arguments) -> Result<Value> {
         .cloned()
 }
 
-/// A wrapper around [`parse_duration`] that converts errors into [`ExecutionError`].
-/// and only returns the duration, rather than returning the remaining input.
-fn _duration(i: &str) -> Result<Duration> {
-    let (_, duration) =
-        parse_duration(i).map_err(|e| ExecutionError::function_error("duration", e.to_string()))?;
-    Ok(duration)
-}
-
-fn _timestamp(i: &str) -> Result<DateTime<FixedOffset>> {
-    DateTime::parse_from_rfc3339(i)
-        .map_err(|e| ExecutionError::function_error("timestamp", e.to_string()))
-}
-
 #[cfg(test)]
 mod tests {
     use crate::context::Context;
     use crate::testing::test_script;
+    #[cfg(feature = "regex")]
+    use crate::ExecutionError::FunctionError;
+    #[cfg(feature = "chrono")]
     use crate::{Program, Value};
+    #[cfg(feature = "chrono")]
     use chrono::{DateTime, FixedOffset};
     use std::collections::HashMap;
 
@@ -662,27 +676,6 @@ mod tests {
     }
 
     #[test]
-    fn test_duration() {
-        [
-            ("duration equal 1", "duration('1s') == duration('1000ms')"),
-            ("duration equal 2", "duration('1m') == duration('60s')"),
-            ("duration equal 3", "duration('1h') == duration('60m')"),
-            ("duration comparison 1", "duration('1m') > duration('1s')"),
-            ("duration comparison 2", "duration('1m') < duration('1h')"),
-            (
-                "duration subtraction",
-                "duration('1h') - duration('1m') == duration('59m')",
-            ),
-            (
-                "duration addition",
-                "duration('1h') + duration('1m') == duration('1h1m')",
-            ),
-        ]
-        .iter()
-        .for_each(assert_script);
-    }
-
-    #[test]
     fn test_starts_with() {
         [
             ("starts with true", "'foobar'.startsWith('foo') == true"),
@@ -702,6 +695,7 @@ mod tests {
         .for_each(assert_script);
     }
 
+    #[cfg(feature = "chrono")]
     #[test]
     fn test_timestamp() {
         [(
@@ -732,6 +726,29 @@ mod tests {
         .for_each(assert_script);
     }
 
+    #[cfg(feature = "chrono")]
+    #[test]
+    fn test_duration() {
+        [
+            ("duration equal 1", "duration('1s') == duration('1000ms')"),
+            ("duration equal 2", "duration('1m') == duration('60s')"),
+            ("duration equal 3", "duration('1h') == duration('60m')"),
+            ("duration comparison 1", "duration('1m') > duration('1s')"),
+            ("duration comparison 2", "duration('1m') < duration('1h')"),
+            (
+                "duration subtraction",
+                "duration('1h') - duration('1m') == duration('59m')",
+            ),
+            (
+                "duration addition",
+                "duration('1h') + duration('1m') == duration('1h1m')",
+            ),
+        ]
+        .iter()
+        .for_each(assert_script);
+    }
+
+    #[cfg(feature = "chrono")]
     #[test]
     fn test_timestamp_variable() {
         let mut context = Context::default();
@@ -744,14 +761,72 @@ mod tests {
         assert_eq!(result, true.into());
     }
 
+    #[cfg(feature = "chrono")]
     #[test]
-    fn test_string() {
+    fn test_chrono_string() {
         [
             ("duration", "duration('1h30m').string() == '1h30m0s'"),
             (
                 "timestamp",
                 "timestamp('2023-05-29T00:00:00Z').string() == '2023-05-29T00:00:00+00:00'",
             ),
+        ]
+        .iter()
+        .for_each(assert_script);
+    }
+
+    #[test]
+    fn test_contains() {
+        let tests = vec![
+            ("list", "[1, 2, 3].contains(3) == true"),
+            ("map", "{1: true, 2: true, 3: true}.contains(3) == true"),
+            ("string", "'foobar'.contains('bar') == true"),
+            ("bytes", "b'foobar'.contains(b'o') == true"),
+        ];
+
+        for (name, script) in tests {
+            assert_eq!(test_script(script, None), Ok(true.into()), "{}", name);
+        }
+    }
+
+    #[cfg(feature = "regex")]
+    #[test]
+    fn test_matches() {
+        let tests = vec![
+            ("string", "'foobar'.matches('^[a-zA-Z]*$') == true"),
+            (
+                "map",
+                "{'1': 'abc', '2': 'def', '3': 'ghi'}.all(key, key.matches('^[a-zA-Z]*$')) == false",
+            ),
+        ];
+
+        for (name, script) in tests {
+            assert_eq!(
+                test_script(script, None),
+                Ok(true.into()),
+                ".matches failed for '{name}'"
+            );
+        }
+    }
+
+    #[cfg(feature = "regex")]
+    #[test]
+    fn test_matches_err() {
+        assert_eq!(
+            test_script(
+                "'foobar'.matches('(foo') == true", None),
+            Err(
+                FunctionError {
+                    function: "matches".to_string(),
+                    message: "'(foo' not a valid regex:\nregex parse error:\n    (foo\n    ^\nerror: unclosed group".to_string()
+                }
+            )
+        );
+    }
+
+    #[test]
+    fn test_string() {
+        [
             ("string", "'foo'.string() == 'foo'"),
             ("int", "10.string() == '10'"),
             ("float", "10.5.string() == '10.5'"),
